co_mbr.sh: Added simple script for a common use-case
'git checkout' branch specified by manifest

That would be the stupid /remotes/<remote>/* branches that are really just
copies of contents of corresponding refs/heads/* on the remote, which are
dangerours to work at. But as Repo continously automerges real branches
when 'repo sync', that is even worse as you don't know what you have.

Category: development